fix(ffi/lua): `LuauLoad` segfault due to use-after-free
We were defering a free of the bytecode too early. `luau_load` expects
the bytecode to be held for as long as it is executing. Also handled ahe
case for if the size of the bytecode is zero, for which case we send a
NULL pointer to `luau_load` instead.