chore(build): replace `go` build wrapper with `go:generate` attrs
This is a follow up inspired by some changes in
https://github.com/CompeyDev/lei/commit/19f3c4f8a09ee6deb00eeedeb115983e360e0d1e,
which was the first commit to start using `go:generate` attributes,
specific for cgo header generation.

Instead of the `build` package being a CLI wrapper around the `go`
command, compiling and injecting required shared libraries to link, we
now have a thin command which compiles Luau CMake subprojects using
Ninja.

We refer to this command in `go:generate` attributes along with the
required static libraries that the file expects. This means that users
must run the following command to initialize the workspace:

```
go generate ./...
```

FUTURE: Might be worth not having a hard dependency on Ninja, or
allowing users to specify their own CMake generator